Refactor: give the runtime the Graph recorder pool
The eight threads that record Graph bodies were created by
GraphAsyncRecordingState, a function-local static in orchestration_api.h — so
one pool lived in each orchestration .so. Those are dlopen'd one per callable
with RTLD_LOCAL and are not released as cases finish, so a process held one
pool, eight threads and their recording storage per registered callable.
Measured on a2a3: 3 orchestration images mapped at once over a four-case pytest
session and 5 over the host_build_graph corpus, i.e. 24-40 recorder threads and
127-212 MB of recorder storage where 8 threads and ~42 MB suffice.

The pool is now the runtime's, in host/graph_recorder_pool.{h,cpp} — the host
target only, since the aicore and aicpu targets compile orchestration/ too and
must not instantiate host threads. One pool per process: counting threads named
hbg-recorder over the same four cases gives 8, independent of how many callables
are registered.

The orchestration .so reaches it through two new ops entries. `job` is a
std::function<void(GraphTaskArgs &)> * the pool moves out of, whether or not it
queues it, since start() takes the callable before it checks capacity — so the
caller treats it as spent on return. Nothing is owned across the .so boundary
either way, and rt_graph_submit's synchronous fallback re-runs its own copy of
the body rather than the job. The AICPU build links weak fallbacks in
runtime_core.cpp — a refusing start and a no-op wait — the same host/device fork
host_tensor_read already uses, which is what makes the device path record
synchronously.

graph_recording_stand_up() catches std::bad_alloc: the node tensor pool is a
nothrow new but the flat arrays are vectors whose resize/reserve throw, and this
now runs on a recorder worker as it starts, where an escaping exception
terminates the process instead of letting prewarm() report the failure. The
temporary orchestration image is also unlinked as soon as dlopen maps it rather
than after the prewarm, so no failure return leaves one behind in /tmp.

Two properties this relies on, both already true:

  - No job outlives the bind that queued it. rt_orchestration_done() ->
    rt_graph_commit() -> graph_record_wait() drains the pool at the end of every
    orchestration, so a job's code cannot still be queued when
    unregister_callable dlcloses the .so it lives in.
  - The runtime a job binds to is a plain global in its own .so
    (orchestration/common.cpp, deliberately not thread_local), so a worker
    shared across callables reads the right one: the job's own inlined code
    reads its own .so's global.

With the pool inside the runtime, standing each worker's recording storage up is
a plain call rather than a symbol handed across the boundary. That deletes the
framework_prewarm_graph_recorders and framework_set_recorder_storage_init
exports, their dlsyms, the OrchestrationStorageInitFunc typedef and the failure
counter that existed only because the count could cross where a return value
could not: registration now calls graph_recorder_prewarm() and reads its bool.
host_orchestration_support/ goes with them.

Cold binds keep what that bought: host_orch minflt per rank was 1220 / 1162 and
1234 / 1189 with the storage stood up lazily inside the first bind, and 1020 /
989 once a worker does it as it starts.

The workers are also named hbg-recorder, so eight mostly-idle threads are
identifiable in top -H or a debugger rather than looking like a leak — which is
what made the before/after above measurable at all.

test_host_orchestration_is_a_self_contained_log_consumer asserted that the
orchestration .so compiles graph_recorder_prewarm.cpp. That file is gone, and the
property is now the opposite one worth guarding: these sources are shared with
the aicore and aicpu targets, so the orchestration .so must compile nothing that
creates host threads. The parameter that carried the old expectation went with
it, since it would have been false for every case.

The pool's own cpput case now drives it through the ops table with a file-local
pool, since the process-wide one lives in a translation unit it does not link;
a weak graph_recorder_stand_up_storage in test_stubs.cpp keeps it from linking
the orchestrator, the same shape as the bind_callable_to_runtime_impl stub. Its
commit_calls counter becomes atomic, because the recorded body calls
rt_graph_commit on a worker while the main thread calls it too.

Verified: cpput 119/119; a2a3 onboard sweep 149 passed, 1 skipped; a2a3sim 60,
a5sim 53; the two a2a3 onboard Graph-recording scene tests golden-checked on
device; graph_execution at --rounds 3 passing its golden on every invocation.
