+++ src/a2a3/runtime/host_build_graph/orchestration/orchestration_api.h
@@ -122,6 +122,16 @@ typedef struct RuntimeOps {
     // the two must stay in lockstep field for field, since the runtime fills the table
     // and this .so calls through it.
     void (*record_orch_phase)(uint32_t kind, uint64_t start_ns, uint64_t end_ns, uint64_t detail);
+    // Queue one Graph body for asynchronous recording, and drain every queued one.
+    // `job` is a `std::function<void(GraphTaskArgs &)> *` the pool moves out of --
+    // whether or not it queues it, since start() takes the callable before it checks
+    // capacity -- so the caller must not invoke it afterwards. Nothing is owned across
+    // the boundary either way: the caller's std::function destructs normally, empty or
+    // not, and rt_graph_submit's fallback re-runs its own copy of the body. The pool is
+    // runtime-owned (host/graph_recorder_pool.h) and the AICPU build links a refusing
+    // fallback, which is what makes the device path record synchronously.
+    bool (*graph_record_start)(RuntimeContext *rt, const GraphTaskArgs &args, void *job);
+    void (*graph_record_wait)(RuntimeContext *rt);
 } RuntimeOps;
 
 /**
@@ -136,244 +146,6 @@ struct RuntimeContext {
     const RuntimeOps *ops;
     ScopeMode pending_scope_mode;
 };
-
-class GraphOwnedArgs {
-public:
-    GraphOwnedArgs() { std::memset(tensors_.data(), 0, sizeof(tensors_)); }
-
-    // The arrays below are sized to the Graph boundary's own capacity, so a
-    // source GraphTaskArgs cannot report more args than they hold and the copy
-    // loops need no runtime bound.
-    void assign(const GraphTaskArgs &source) {
-        args_.reset();
-        for (int32_t i = 0; i < source.tensor_count(); ++i) {
-            tensors_[static_cast<size_t>(i)].copy(source.tensor(i).ref());
-            switch (source.tag(i)) {
-            case TensorArgType::INPUT:
-                args_.add_input(tensors_[static_cast<size_t>(i)]);
-                break;
-            case TensorArgType::OUTPUT_EXISTING:
-                args_.add_output(tensors_[static_cast<size_t>(i)]);
-                break;
-            case TensorArgType::INOUT:
-                args_.add_inout(tensors_[static_cast<size_t>(i)]);
-                break;
-            case TensorArgType::NO_DEP:
-                args_.add_no_dep(tensors_[static_cast<size_t>(i)]);
-                break;
-            case TensorArgType::OUTPUT:
-                args_.set_error("Runtime-allocated output is not supported at a Graph boundary");
-                break;
-            }
-        }
-        for (int32_t i = 0; i < source.scalar_count(); ++i) {
-            scalars_[static_cast<size_t>(i)] = source.scalar(i);
-            args_.add_scalar(scalars_[static_cast<size_t>(i)]);
-        }
-        args_.launch_spec = source.launch_spec;
-        args_.set_allow_early_resolve(source.allow_early_resolve());
-        if (source.task_timing_slot() != TASK_TIMING_SLOT_NONE) {
-            args_.set_task_timing_slot(source.task_timing_slot());
-        }
-        args_.set_predicate(source.predicate());
-    }
-
-    GraphTaskArgs &args() { return args_; }
-
-private:
-    std::array<simpler::hbg::Tensor, GRAPH_MAX_TENSOR_ARGS> tensors_{};
-    std::array<uint64_t, GRAPH_MAX_SCALAR_ARGS> scalars_{};
-    GraphTaskArgs args_;
-};
-
-class GraphAsyncRecordingState {
-public:
-    GraphAsyncRecordingState() {
-        for (size_t i = 0; i < kJobCapacity; ++i) {
-            free_owned_args_[i] = kJobCapacity - i - 1;
-        }
-    }
-    ~GraphAsyncRecordingState() { shutdown(); }
-
-    GraphAsyncRecordingState(const GraphAsyncRecordingState &) = delete;
-    GraphAsyncRecordingState &operator=(const GraphAsyncRecordingState &) = delete;
-
-    // A workload that cuts its forward pass into up to eight Definitions should
-    // not create pthreads between outer shell submissions. Callable registration
-    // parks those workers before the first run; start() can still grow to the
-    // Definition limit if a workload records more identities concurrently.
-    bool prewarm() {
-        std::unique_lock<std::mutex> lock(mutex_);
-        if (stopping_) return false;
-        while (workers_.size() < kPrewarmedWorkerCount) {
-            if (!create_worker_locked()) break;
-        }
-        const size_t target = workers_.size();
-        cv_.wait(lock, [&]() {
-            return ready_workers_ >= target || stopping_;
-        });
-        return !stopping_ && target == kPrewarmedWorkerCount;
-    }
-
-    template <typename Job>
-    bool start(const GraphTaskArgs &args, Job &&job) {
-        std::function<void(GraphTaskArgs &)> next;
-        try {
-            next = std::forward<Job>(job);
-        } catch (...) {
-            return false;
-        }
-
-        size_t owned_args_index;
-        {
-            std::lock_guard<std::mutex> lock(mutex_);
-            if (stopping_ || free_owned_args_count_ == 0 || job_count_ == kJobCapacity) return false;
-            owned_args_index = free_owned_args_[--free_owned_args_count_];
-        }
-        owned_args_[owned_args_index].assign(args);
-
-        std::unique_lock<std::mutex> lock(mutex_);
-        if (stopping_) {
-            free_owned_args_[free_owned_args_count_++] = owned_args_index;
-            return false;
-        }
-        PendingJob &pending = jobs_[job_tail_];
-        pending.function = std::move(next);
-        pending.owned_args_index = owned_args_index;
-        job_tail_ = (job_tail_ + 1) % kJobCapacity;
-        job_count_++;
-        const size_t desired_workers = std::min(kMaxWorkerCount, job_count_ + active_jobs_);
-        while (workers_.size() < desired_workers) {
-            if (!create_worker_locked()) break;
-        }
-        if (workers_.empty()) {
-            job_tail_ = (job_tail_ + kJobCapacity - 1) % kJobCapacity;
-            PendingJob &rollback = jobs_[job_tail_];
-            rollback.function = {};
-            job_count_--;
-            free_owned_args_[free_owned_args_count_++] = rollback.owned_args_index;
-            return false;
-        }
-        lock.unlock();
-        cv_.notify_one();
-        // graph_begin() has already installed the keyed in-flight entry and
-        // submitted the zero-heap outer shell. Enqueuing the private job is
-        // therefore the last dependency of the caller; graph_prepare() and all
-        // node recording may start after later shells are submitted.
-        return true;
-    }
-
-    // Wait for every queued and running recording. A recording thread returns
-    // immediately: it may reach this through rt_orchestration_done in a Graph body
-    // and must never wait for its own job, nor for a sibling's — the sibling makes
-    // progress independently and waiting on it would trade a recording thread for
-    // nothing.
-    void wait() {
-        std::unique_lock<std::mutex> lock(mutex_);
-        if (is_worker_thread_locked(std::this_thread::get_id())) return;
-        cv_.wait(lock, [&]() {
-            return job_count_ == 0 && active_jobs_ == 0;
-        });
-    }
-
-private:
-    static constexpr size_t kPrewarmedWorkerCount = 8;
-    static constexpr size_t kMaxWorkerCount = GRAPH_MAX_DEFINITIONS;
-    static constexpr size_t kJobCapacity = GRAPH_MAX_DEFINITIONS;
-
-    struct PendingJob {
-        std::function<void(GraphTaskArgs &)> function;
-        size_t owned_args_index{0};
-    };
-
-    bool is_worker_thread_locked(std::thread::id id) const {
-        for (const std::thread &worker : workers_) {
-            if (worker.get_id() == id) return true;
-        }
-        return false;
-    }
-
-    bool create_worker_locked() {
-        if (stopping_ || workers_.size() >= kMaxWorkerCount) return false;
-        try {
-            workers_.emplace_back([this]() {
-                run();
-            });
-            return true;
-        } catch (...) {
-            return false;
-        }
-    }
-
-    void run() {
-        {
-            std::lock_guard<std::mutex> lock(mutex_);
-            ready_workers_++;
-        }
-        cv_.notify_all();
-        for (;;) {
-            PendingJob current;
-            {
-                std::unique_lock<std::mutex> lock(mutex_);
-                cv_.wait(lock, [&]() {
-                    return job_count_ != 0 || stopping_;
-                });
-                if (stopping_ && job_count_ == 0) return;
-                PendingJob &pending = jobs_[job_head_];
-                current.function = std::move(pending.function);
-                current.owned_args_index = pending.owned_args_index;
-                job_head_ = (job_head_ + 1) % kJobCapacity;
-                job_count_--;
-                active_jobs_++;
-            }
-            current.function(owned_args_[current.owned_args_index].args());
-            current.function = {};
-            {
-                std::lock_guard<std::mutex> lock(mutex_);
-                free_owned_args_[free_owned_args_count_++] = current.owned_args_index;
-                active_jobs_--;
-            }
-            cv_.notify_all();
-        }
-    }
-
-    void shutdown() {
-        wait();
-        {
-            std::lock_guard<std::mutex> lock(mutex_);
-            stopping_ = true;
-        }
-        cv_.notify_all();
-        for (std::thread &worker : workers_) {
-            if (worker.joinable()) worker.join();
-        }
-    }
-
-    std::vector<std::thread> workers_;
-    std::mutex mutex_;
-    std::condition_variable cv_;
-    std::array<GraphOwnedArgs, kJobCapacity> owned_args_;
-    std::array<size_t, kJobCapacity> free_owned_args_{};
-    std::array<PendingJob, kJobCapacity> jobs_;
-    size_t free_owned_args_count_{kJobCapacity};
-    size_t job_head_{0};
-    size_t job_tail_{0};
-    size_t job_count_{0};
-    size_t ready_workers_{0};
-    size_t active_jobs_{0};
-    bool stopping_{false};
-};
-
-// External linkage on purpose: `static inline` would give every translation
-// unit that submits a Graph its own recorder and its own worker threads, so a
-// commit reached from one TU would not wait for a recording another TU started.
-// Vague linkage keeps one instance — and one pool — per loaded SO.
-inline GraphAsyncRecordingState &rt_graph_async_recording() {
-    // Keep the bounded pool alive across runs so steady-state misses pay only a
-    // condition-variable wakeup. The SO's destructor joins it before dlclose.
-    static GraphAsyncRecordingState state;
-    return state;
-}
 
 // =============================================================================
 // Inline Convenience Wrappers (call through ops table)
@@ -531,10 +303,9 @@ static inline bool rt_graph_end() {
 }
 
 static inline void rt_graph_commit() {
-    GraphAsyncRecordingState &async = rt_graph_async_recording();
-    async.wait();
-
     RuntimeContext *rt = current_runtime();
+    if (rt->ops->graph_record_wait != nullptr) rt->ops->graph_record_wait(rt);
+
     if (!rt->ops->is_fatal(rt) && rt->ops->graph_commit != nullptr) rt->ops->graph_commit(rt);
 }
 
@@ -763,9 +534,10 @@ static inline GraphSubmitResult rt_submit_graph_impl(uint64_t graph_key, const G
     GraphScopeResult result = rt_graph_begin(graph_key, args);
     const uint64_t _begun_ns = rt_orch_phase_now_ns();
     if (result.recording) {
-        GraphAsyncRecordingState &async = rt_graph_async_recording();
         void *handle = result.recording_handle;
-        auto record = [invoke, handle](GraphTaskArgs &record_args) mutable {
+        // A std::function rather than a bare lambda because the pool takes it as one
+        // through the ops table's void *.
+        std::function<void(GraphTaskArgs &)> job = [invoke, handle](GraphTaskArgs &record_args) mutable {
             try {
                 if (!rt_graph_prepare(handle, record_args)) {
                     rt_graph_abort(handle);
@@ -777,7 +549,14 @@ static inline GraphSubmitResult rt_submit_graph_impl(uint64_t graph_key, const G
                 rt_graph_abort(handle);
             }
         };
-        if (!async.start(args, std::move(record))) {
+        // The pool takes the callable whether or not it queues it: start() moves it into
+        // its own storage before it checks capacity, so `job` is empty either way. That
+        // costs nothing here, because the fallback below re-runs `invoke` -- captured by
+        // value, so unaffected -- rather than the job.
+        RuntimeContext *record_rt = current_runtime();
+        const bool queued =
+            record_rt->ops->graph_record_start != nullptr && record_rt->ops->graph_record_start(record_rt, args, &job);
+        if (!queued) {
             try {
                 if (!rt_graph_prepare(handle, args)) {
                     rt_graph_abort(handle);
