+++ tests/ut/cpp/stubs/test_stubs.cpp
@@ -133,6 +133,16 @@ uint64_t get_test_reg_stub_value() { return g_test_reg; }
 uint64_t get_test_reg_stub_base_addr() { return g_test_reg_base_addr; }
 
 // =============================================================================
+// orchestrator.cpp stub (graph_recorder_stand_up_storage)
+// =============================================================================
+
+// A recorder worker stands its recording storage up as it starts
+// (host/graph_recorder_pool.h). The real one is in orchestrator.cpp, which the pool's own
+// threading test does not link -- and does not need, since nothing it records reaches the
+// storage. Weak so a test that links the orchestrator gets the real one.
+__attribute__((weak)) bool graph_recorder_stand_up_storage() { return true; }
+
+// =============================================================================
 // runtime_maker.cpp stub (bind_callable_to_runtime_impl)
 // =============================================================================
